Move Sync Dev Settings to live in an internal-only module
