+++ sync/sync-impl/src/main/AndroidManifest.xml
@@ -20,13 +20,6 @@
 
     <application>
         <activity
-            android:name=".ui.SyncInternalSettingsActivity"
-            android:exported="false"
-            android:label="@string/sync_internal_activity_label"
-            android:screenOrientation="portrait"
-            android:parentActivityName="com.duckduckgo.app.settings.SettingsActivity"/>
-
-        <activity
             android:name=".ui.SyncActivity"
             android:exported="false"
             android:label="@string/sync_screen_title"
