+++ sync/sync-internal/src/main/res/layout/activity_internal_sync_settings.xml
