+++ sync/sync-internal/src/main/res/layout/item_connected_device.xml
