+++ sync/sync-internal/src/main/res/values/donottranslate.xml
@@ -18,4 +18,34 @@
     <!-- sync dev settings -->
     <string name="syncSettingsTitle">Sync Dev Settings</string>
     <string name="syncSettingsSubtitle">Sync Dev Internal Test</string>
+
+    <!-- Internal Setting -->
+    <string name="sync_internal_activity_label">Sync Setup Internal</string>
+    <string name="sync_internal_environment_header">Environment Settings</string>
+    <string name="sync_internal_environment_title">Use Dev Environment?</string>
+    <string name="sync_internal_environment_subtitle_prod">https://sync.duckduckgo.com</string>
+    <string name="sync_internal_environment_subtitle_dev">https://sync-staging.duckduckgo.com</string>
+    <string name="sync_internal_account_header">Account Settings</string>
+    <string name="sync_internal_device_header">Device Settings</string>
+    <string name="sync_internal_recovery_code">Recovery code</string>
+    <string name="sync_internal_recovery_code_cta">Copy Recovery code</string>
+    <string name="sync_internal_favicons_prompt_cta">Reset Favicons Prompt</string>
+    <string name="userIdSectionHeader">User ID</string>
+    <string name="deviceIdSectionHeader">Device Id</string>
+    <string name="deviceNameSectionHeader">Device Name</string>
+    <string name="sync_internal_promotions_header">Sync Promotions</string>
+    <string name="sync_internal_promotions_clear_history_bookmark_added_promo">Clear history (bookmark added dialog)</string>
+    <string name="sync_internal_promotions_clear_history_bookmark_screen_promo">Clear history (bookmark screen)</string>
+    <string name="sync_internal_promotions_clear_history_password_screen_promo">Clear history (password screen)</string>
+
+    <string name="sync_internal_launch_recover_data_screen">Launch Recover Data Screen</string>
+    <string name="sync_internal_block_store_write_recovery_code">Save account recovery code &amp; device ID to Block Store</string>
+
+    <!-- Block Store -->
+    <string name="sync_internal_block_store_header">Persistent Storage / Block Store</string>
+    <string name="sync_internal_block_store_current_value_label">Current stored value (SyncRecovery key)</string>
+    <string name="sync_internal_block_store_recovery_code_hint">Recovery code</string>
+    <string name="sync_internal_block_store_device_id_hint">Device ID (optional, 8–64 chars)</string>
+    <string name="sync_internal_block_store_write">Write</string>
+    <string name="sync_internal_block_store_clear">Clear</string>
 </resources>
